refactor: add transformer factory
