+++ support/transformers/SecondSourceTransformer.js
@@ -1,0 +1,6 @@
+function SecondSourceTransformer(html, link) {
+  console.log("SecondSourceTransformer", html, link);
+  return []; // events array
+}
+
+module.exports = { SecondSourceTransformer };
