+++ support/transformers/transformerFactory.js
@@ -1,0 +1,15 @@
+const { ChooseChicagoTransformer } = require("./ChooseChicagoTransformer.js");
+const { SecondSourceTransformer } = require("./SecondSourceTransformer.js");
+
+const providersMap = {
+  'CHOOSECHICAGO': ChooseChicagoTransformer,
+  'SECONDSOURCE': SecondSourceTransformer
+}
+
+class TransformerFactory {
+  getTransformer(provider){
+    return providersMap[provider];
+  }
+}
+
+module.exports = { TransformerFactory }
